Update TLS certificate instructions
mozroots is deprecated and replaced by cert-sync

For some reason, mono still rejects certificate chains with expired
roots, even if the certificate has a valid chain to another root. This
was fixed in OpenSSL 1.1.0 (we have 1.1.1d), but for some reason it
doesn't work in Mono. The easiest fix is to deselect the expired
DST_Root_CA_X3 in /etc/ca-certificates.conf before exporting the
certificates to the Mono trust store.
NOTE: I don't know if this file is overwritten on package upgrade. I
suppose we'll find out eventually.