docs: Fix broken grafana-labs URL
While the "cloud-vps-project-board" dashboard is still marked
as featured on the Grafana labs dashboard, and still shows current
metrics for pre-2019 instances, it does not appear to know about
any instances created this year.

From searching in grafana-labs database, it seems there is a newer
dashboard that is probably meant to replace it:

https://grafana-labs.wikimedia.org/d/000000059/cloud-vps-project-board?var-project=cvn

https://grafana-labs.wikimedia.org/d/8Npp-46Zz/project-overview?orgId=1&var-project=cvn

Follow the "VM instance" links to see metrics about individual
instances, e.g.
https://grafana-labs.wikimedia.org/d/000000590/instance-details?%3ForgId=1&var-project=cvn&var-job=node&var-node=cvn-app10&orgId=1&from=now-24h&to=now

Change-Id: I7c8e8212ca4694e04c39b7616d0f41795275aed9