megatable: Upgrade jQuery from 2.1.0 to 3.7.1
This is in prep for the next patch, which changes an $.ajax() call
to native Fetch. This upgrade is important for that because in
jQuery 2.x, the $.when() function only resolved jQuery.Deferred values,
not native Promise.

In jQuery 3.x, the $.when() function changed to also resolve native
Promise and other Thenable objects, such as those returned by fetch().

Change-Id: Ie9a5889a053ed16aa4eb3ba130ec961034930ca7