megatable: Fix "XML Parsing Error: not well-formed" warning
Even though the tool works fine, when we fetch the `large.dblist` text
file, the native XMLHttpRequest behind $.ajax() is attempting to also
prepare an XML Document object in case we want it. This fails of course
because it is a text file, not an XML file. The error is not an exception,
but an error event logged directly to the console, so the tool works fine.

```
XML Parsing Error: not well-formed
Location: https://cvn.wmcloud.org/megatable/var/large.dblist
Line Number 1, Column 2:
```

Avoid the error by using fetch() and loading the response in text()
mode instead.

Change-Id: I2bd82e3f1c107591455e2bf6880daf8a496cff6b